added dislikes,likes

+++ pkg/posts.go
@@ -16,24 +16,20 @@ type PostModel struct {
 func (m *PostModel) Latest() ([]*Post, error) {
 	stmt := `SELECT * FROM post ORDER BY created_at ASC LIMIT 200`
 	rows, err := m.DB.Query(stmt)
-	if err != nil {
-		return nil, err
-	}
+
 	defer rows.Close()
 	var posts []*Post
 	for rows.Next() {
 		s := &Post{}
 		err = rows.Scan(&s.ID, &s.Title, &s.Content, &s.User_id, &s.Created_at)
 		s.Category_name, err = m.getCategoryRelation(&s.ID)
-		s.Like, err = m.getCountLikesByPostId(&s.ID)
+		s.Like, s.Dislike, err = m.getCountLikesByPostId(&s.ID)
 		if err != nil {
 			return nil, err
 		}
 		posts = append(posts, s)
 	}
-	if err = rows.Err(); err != nil {
-		return nil, err
-	}
+
 	return posts, nil
 }
 
@@ -84,27 +80,36 @@ func (m *PostModel) GetPost(id int) (*Post, error) {
 	return post, nil
 }
 
-func (m *PostModel) getCountLikesByPostId(id *int) (int, error) {
+func (m *PostModel) getCountLikesByPostId(id *int) (int, int, error) {
 	var id_post int
-	stmt := `SELECT id FROM like WHERE post_id = ?`
+	stmt := `SELECT id FROM like WHERE post_id = ? AND type_of = "like"`
 	rows, err := m.DB.Query(stmt, id)
 	if err != nil {
-		return 0, err
+		return 0, 0, err
 	}
 	defer rows.Close()
 	var likes []int
+	var dislikes []int
+	stmt_dislike := `SELECT id FROM like WHERE post_id = ? AND type_of = "dislike"`
+	rows_dis, err := m.DB.Query(stmt_dislike, id)
+	for rows_dis.Next() {
+		var c int
+		err = rows_dis.Scan(&id_post)
+		dislikes = append(dislikes, c)
+	}
 	for rows.Next() {
 		var s int
 		err = rows.Scan(&id_post)
 		if err != nil {
-			return 0, err
+			return 0, 0, err
 		}
 		likes = append(likes, s)
 	}
 	if err = rows.Err(); err != nil {
-		return 0, err
+		return 0, 0, err
 	}
-	return len(likes), nil
+
+	return len(likes), len(dislikes), nil
 }
 
 func (m *PostModel) Insert(header, description string, userId int, created_at time.Time, category []string) error {
@@ -112,9 +117,6 @@ func (m *PostModel) Insert(header, description string, userId int, created_at ti
     VALUES(?,?,?, current_date)`
 
 	result, err := m.DB.Exec(stmt, header, description, userId, created_at)
-	if err != nil {
-		return err
-	}
 
 	id, err := result.LastInsertId()
 	for _, category_id := range category {
@@ -128,7 +130,6 @@ func (m *PostModel) Insert(header, description string, userId int, created_at ti
 	if err != nil {
 		return err
 	}
-	log.Println(id)
 
 	return nil
 }
@@ -234,28 +235,34 @@ func getPostsByIDList(db *sql.DB, idList []int) ([]*Post, error) {
 	return posts, nil
 }
 
-func (m *PostModel) CreateLike(user_id, post_id int) error {
+func (m *PostModel) CreateLike(user_id, post_id int, type_of_like string) error {
 	likedOrNot := m.checkLike(user_id, post_id)
 	switch likedOrNot {
 	case true:
-		stmt := "DELETE FROM like WHERE user_id = ? AND post_id = ?"
-		_, err := m.DB.Exec(stmt, user_id, post_id)
+		stmt := "DELETE FROM like WHERE user_id = ? AND post_id = ? AND type_of = ?"
+		result, err := m.DB.Exec(stmt, user_id, post_id, type_of_like)
+		res, err := result.RowsAffected()
+		if res == int64(0) {
+			stmt := "DELETE FROM like WHERE user_id = ? AND post_id = ?"
+			_, err := m.DB.Exec(stmt, user_id, post_id, type_of_like)
+			if err != nil {
+
+				return err
+			}
+			m.CreateLike(user_id, post_id, type_of_like)
+		}
 		if err != nil {
 			return err
 		}
 	case false:
-		stmt := `INSERT INTO like (user_id, post_id)
-    VALUES(?,?)`
-		result, err := m.DB.Exec(stmt, user_id, post_id)
+		stmt := `INSERT INTO like(post_id, user_id,type_of)
+    VALUES(?,?,?)`
+		_, err := m.DB.Exec(stmt, post_id, user_id, type_of_like)
+
 		if err != nil {
 			return err
 		}
-		id, err := result.LastInsertId()
-		if err != nil {
-			return err
-		}
-		log.Println(id)
-		return nil
+
 	}
 	return nil
 }
